Better implementation of raire service status checking.
